+++ packages/jest-expo/src/preset/moduleMocks/expoModules.js
@@ -568,6 +568,7 @@ module.exports = {
           },
         ],
         ExpoRouterNativeLinkPreview: [],
+        ExpoRouter: [],
         ExpoScreenCapture: [
           { name: 'allowScreenCapture', argumentsCount: 0, key: 'allowScreenCapture' },
           {
@@ -1210,6 +1211,10 @@ module.exports = {
           getDevicePushTokenAsync: { type: 'function' },
           removeListeners: { type: 'function' },
           unregisterForNotificationsAsync: { type: 'function' },
+        },
+        ExpoRouter: {
+          Material3Color: { type: 'function' },
+          Material3DynamicColor: { type: 'function' },
         },
         ExpoRouterNativeLinkPreview: {
           addListener: { type: 'function' },